backend/pyproject.toml
[project]
name = "caproject-backend"
version = "0.1.0"
description = "Capstone backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.22.0",
    "pydantic>=1.10.0",
    "pandas",
    "sqlalchemy>=1.4.0",
    "flask>=2.2.0",
    "flask-migrate>=4.1.0",
    "flask-cors==6.0.2",
    "flask-sqlalchemy>=3.1.1",
    "mysql-connector-python>=8.0.0",
    "python-dotenv>=0.20.0",
    "alembic>=1.10.0",
    "requests>=2.25",
    "openpyxl>=3.0",
    "pymysql>=1.0",
    "pyjwt>=2.10.1",
    "werkzeug==3.0.1",
    "gunicorn==21.2.0",
]

[tool.setuptools]
py-modules = ["config", "main", "models"]
